+++ apps/web-roo-code/src/app/evals/types.ts
@@ -1,0 +1,9 @@
+import type { TaskMetrics, Run } from "@roo-code/evals"
+
+export type EvalRun = Run & {
+	label: string
+	score: number
+	languageScores?: Record<"go" | "java" | "javascript" | "python" | "rust", number>
+	taskMetrics: TaskMetrics
+	modelId?: string
+}
